// src/enums/message.ts
// 消息类型
export enum MessageType {
  TEXT = 'text',
  PHOTO = 'photo',
  VIDEO = 'video',
  VOICE = 'voice',
  DOCUMENT = 'document',
  STICKER = 'sticker',
  LOCATION = 'location',
  MENTION = 'mention',
  COMMAND = 'command',
  IMAGE = 'photo',
  AUDIO = 'audio',
  FILE = 'file',
  OTHER = 'other',
  UNKNOWN = '未知消息类型',
}